refactor: switch to trace ray api

+++ src/PlayerData.cs
@@ -42,12 +42,12 @@ public class PlayerData
     public bool SpectatingC4 { get; set; }
 
     [Key(12)]
-    public Dictionary<string, int> OcclusionFraction { get; set; }
+    public Dictionary<string, float> OcclusionFraction { get; set; }
 
     public PlayerData(string steamId, string name)
     {
         this.SteamId = steamId;
         this.Name = name;
-        this.OcclusionFraction = new Dictionary<string, int>();
+        this.OcclusionFraction = new Dictionary<string, float>();
     }
 }
